split read queue into a file
